First draft of workflow launch on release publish
Launch the sl_core workflow on PW when there is a release in this repo.  Note that sl_core is the ML-code repo and this repo is the ML-archive repo - we want retraining to be triggered when new data is available, so it is releases in the ML-archive that launch the training (and not changes to the ML-code repo).

The workflow-parameters will need to be changed to match the specification of ICON-ModEx iterations; the current configuration was copied from sl_core/apirun/main.py@0a0f268b1198412cea98a7b6ab10fd0d3ff3a8d2 which is just temporary for testing. Both the workflow file in this repo and the API run script in sl_core should be updated together.