Do not colocate in sbatch, remove df index
If we use sbatch for colocation, all workers will pull a docker
image.  For general production, this is fine because it will be run
once in a while.  Also, it's a nice way to warm up the cluster
and get the worker nodes spun up before starting the SuperLearner
jobs. But for dev, I could overrun my 100 anonymous pulls
per 6 hours on the cluster if the worker nodes are restarted.