Renamed input files to match preprocessing expectations
We can change the hard coded file names in the preprocessing
scripts to match preferences later.  For now, I prefer to
verify that it was indeed only a file naming issue that
prevented the iteration from completing. I want to preserve
the input data and code as-is, only renaming the files.