Append train set to predict set so consistent predictions for all
The goal is to grab the predicted value and estimated error for all sites - including the sites where we have actual observations. We will be able to do this because the predict set is not randomly split/shuffled so we can always recover the site ID from the .ixy.  We can then look up the observed value in the training set to compare the real error in addition to the estimated error. We can filter out the sites where we have observations based on their site ID, later.