Update data with same NA marker for all missing data
Some places used n/a or N/A as well as NA. This does not
confuse Pandas, but it does make it hard to use command
line tools (grep, awk, sed) to quickly grab info about the data.
There should be no change to the information content of the data.